+++ data/neoforge/src/main/java/earth/terrarium/common_storage_lib/data/sync/AttachmentData.java
@@ -29,7 +29,7 @@ public record AttachmentData<T>(DataSyncSerializer<T> serializer, @Nullable T da
             buf.writeBoolean(false);
         } else {
             buf.writeBoolean(true);
-            serializer.getCodec().encode(buf, (T) data);
+            serializer.getCodec().encode(buf, data);
         }
     }