+++ changelog.md
@@ -1,2 +1,1 @@
-- Fix lookups not being registered properly on Fabric
-- Fix class loading lockup with fluid ingredients
+- Added syncing to level data (NeoForge)